requirements.txt
certifi==2018.10.15
chardet==3.0.4
idna==2.7
pkg-resources==0.0.0
python-decouple==3.1
requests==2.20.1
urllib3==1.24.1
